refactor: move compiler into subfolder
